requirements.txt
# Core RAG Framework
langchain==0.1.0
langchain-community==0.0.13

# Vector Database
chromadb==0.4.22

# Embeddings (FREE)
sentence-transformers==2.3.1

# PDF Processing
pypdf==4.0.1
pdfplumber==0.10.3

# Data Processing
numpy>=1.26.0
pandas>=2.2.0

# Evaluation
ragas==0.1.1

# UI
streamlit==1.30.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1